refactor(gitops): split mod.rs (956 lines) into submodules
U-16 (<=800 lines/file) was violated. Move git invocation helpers,
repo state, index snapshot, commit/staging, remote, URL validation,
and object hashing into focused submodules. All pub fn signatures
unchanged; history.rs and history_impl.rs untouched.

- mod.rs        360 (was 956)
- exec.rs       135
- repo.rs        82
- index.rs      103
- commit.rs      63
- remote.rs     150
- url.rs         94
- objects.rs     52
