+++ src/gitops/mod.rs
@@ -1,16 +1,36 @@
 mod history;
 mod history_impl;
 
+mod commit;
+mod exec;
+mod index;
+mod objects;
+mod remote;
+mod repo;
+mod url;
+
 pub use history::*;
 
-use std::fs::{self, OpenOptions};
-use std::io::Write;
-use std::path::{Path, PathBuf};
-use std::process::{Command, Output, Stdio};
+pub use commit::{commit, commit_paths_if_changed, stage_path};
+pub use exec::{run_git, run_git_allow_failure, run_git_allow_failure_restricted};
+pub use index::{IndexSnapshot, restore_index, snapshot_index};
+pub use objects::{diff_path, fsck};
+pub use remote::{
+    ahead_behind_main, ahead_behind_refs, fetch_origin_history_branch_if_present,
+    fetch_origin_main_if_present, pull_rebase_main, push_main_with_tags,
+    remote_exists, remote_tracking_history_exists, remote_tracking_main_exists, remote_url,
+    set_remote_origin,
+};
+pub use repo::{
+    checkout_path_from_ref, create_annotated_tag, ensure_repo_initialized,
+    has_staged_changes_for_path, head, repo_is_initialized, resolve_ref, short_head,
+};
+pub use url::validate_git_url;
 
-use anyhow::{Context, Result, anyhow};
-
-use crate::state::AppContext;
+// Re-export internals needed by history_impl (accessed via `super::`)
+pub(crate) use exec::run_git_in_with_env;
+pub(crate) use objects::{TempFile, hash_object_bytes, hash_object_file, read_blob};
+pub(crate) use repo::ensure_local_identity;
 
 pub const HISTORY_BRANCH: &str = "loom-history";
 const HISTORY_BRANCH_REF: &str = "refs/heads/loom-history";
@@ -23,628 +43,12 @@ const HISTORY_COMPACT_AFTER_SEGMENTS: usize = 8;
 const HISTORY_RETAIN_RECENT_SEGMENTS: usize = 4;
 const HISTORY_RETAIN_ARCHIVES: usize = 4;
 
-#[derive(Debug, Clone, Copy)]
-enum GitEnvMode {
-    Normal,
-    Restricted,
-}
-
-fn run_git_raw_in_with_env_and_input(
-    repo_dir: &Path,
-    envs: &[(&str, &str)],
-    input: Option<&[u8]>,
-    args: &[&str],
-) -> Result<Output> {
-    run_git_raw_in_with_env_mode_and_input(repo_dir, envs, input, args, GitEnvMode::Normal)
-}
-
-fn run_git_raw_in_with_env_mode_and_input(
-    repo_dir: &Path,
-    envs: &[(&str, &str)],
-    input: Option<&[u8]>,
-    args: &[&str],
-    env_mode: GitEnvMode,
-) -> Result<Output> {
-    let mut command = Command::new("git");
-    if matches!(env_mode, GitEnvMode::Restricted) {
-        command.env_clear();
-        if let Some(path) = std::env::var_os("PATH") {
-            command.env("PATH", path);
-        }
-        command
-            .env("GIT_CONFIG_NOSYSTEM", "1")
-            .env("GIT_CONFIG_GLOBAL", "/dev/null")
-            .env("GIT_TERMINAL_PROMPT", "0");
-    }
-    command
-        .current_dir(repo_dir)
-        .arg("-c")
-        .arg("commit.gpgsign=false")
-        .arg("-c")
-        .arg("tag.gpgSign=false")
-        .arg("-c")
-        .arg("protocol.allow=never")
-        .arg("-c")
-        .arg("protocol.https.allow=always")
-        .arg("-c")
-        .arg("protocol.ssh.allow=always");
-    if matches!(env_mode, GitEnvMode::Normal) {
-        command.arg("-c").arg("protocol.file.allow=always");
-    }
-    command
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped())
-        .args(args);
-    for (key, value) in envs {
-        command.env(key, value);
-    }
-    if input.is_some() {
-        command.stdin(Stdio::piped());
-    }
-
-    let mut child = command
-        .spawn()
-        .with_context(|| format!("failed to run git {:?}", args))?;
-    if let Some(bytes) = input {
-        let mut stdin = child.stdin.take().context("failed to open git stdin")?;
-        stdin
-            .write_all(bytes)
-            .with_context(|| format!("failed to write git stdin for {:?}", args))?;
-    }
-
-    child
-        .wait_with_output()
-        .with_context(|| format!("failed to read git output for {:?}", args))
-}
-
-pub fn run_git(ctx: &AppContext, args: &[&str]) -> Result<String> {
-    run_git_in(&ctx.root, args)
-}
-
-fn run_git_in(repo_dir: &Path, args: &[&str]) -> Result<String> {
-    run_git_in_with_env(repo_dir, &[], args)
-}
-
-fn run_git_in_with_env(repo_dir: &Path, envs: &[(&str, &str)], args: &[&str]) -> Result<String> {
-    let output = run_git_allow_failure_in_with_env(repo_dir, envs, args)?;
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-        return Err(anyhow!("git {:?} failed: {}", args, stderr));
-    }
-    Ok(String::from_utf8_lossy(&output.stdout).trim().to_string())
-}
-
-fn run_git_in_with_input(repo_dir: &Path, args: &[&str], input: &[u8]) -> Result<String> {
-    let output = run_git_raw_in_with_env_and_input(repo_dir, &[], Some(input), args)?;
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-        return Err(anyhow!("git {:?} failed: {}", args, stderr));
-    }
-    Ok(String::from_utf8_lossy(&output.stdout).trim().to_string())
-}
-
-pub fn run_git_allow_failure(ctx: &AppContext, args: &[&str]) -> Result<Output> {
-    run_git_allow_failure_in(&ctx.root, args)
-}
-
-pub fn run_git_allow_failure_restricted(ctx: &AppContext, args: &[&str]) -> Result<Output> {
-    run_git_raw_in_with_env_mode_and_input(&ctx.root, &[], None, args, GitEnvMode::Restricted)
-}
-
-fn run_git_allow_failure_in(repo_dir: &Path, args: &[&str]) -> Result<Output> {
-    run_git_allow_failure_in_with_env(repo_dir, &[], args)
-}
-
-fn run_git_allow_failure_in_with_env(
-    repo_dir: &Path,
-    envs: &[(&str, &str)],
-    args: &[&str],
-) -> Result<Output> {
-    run_git_raw_in_with_env_and_input(repo_dir, envs, None, args)
-}
-
-pub fn ensure_repo_initialized(ctx: &AppContext) -> Result<()> {
-    let repo_probe = run_git_allow_failure(ctx, &["rev-parse", "--git-dir"])?;
-    if repo_probe.status.success() {
-        ensure_local_identity(ctx)?;
-        return Ok(());
-    }
-    if ctx.root.join(".git").exists() {
-        return Err(anyhow!("git metadata exists but repository is not healthy"));
-    }
-
-    let init_main = run_git_allow_failure(ctx, &["init", "-b", "main"])?;
-    if !init_main.status.success() {
-        run_git(ctx, &["init"])?;
-        let _ = run_git_allow_failure(ctx, &["branch", "-M", "main"])?;
-    }
-
-    ensure_local_identity(ctx)?;
-    Ok(())
-}
-
-pub fn repo_is_initialized(ctx: &AppContext) -> Result<bool> {
-    let repo_probe = run_git_allow_failure(ctx, &["rev-parse", "--git-dir"])?;
-    Ok(repo_probe.status.success())
-}
-
-pub fn has_staged_changes_for_path(ctx: &AppContext, path: &Path) -> Result<bool> {
-    let path_str = path.to_string_lossy();
-    let output = run_git_allow_failure(ctx, &["diff", "--cached", "--quiet", "--", &path_str])?;
-    Ok(!output.status.success())
-}
-
-/// Captured registry index used to restore staging after a failed mutation.
-///
-/// Backs up the underlying active Git index file as a whole, so every per-entry
-/// flag (`skip-worktree`, `assume-unchanged`, intent-to-add, fsmonitor cache)
-/// survives the rollback. The previous `git write-tree`/`read-tree` design
-/// dropped these flags because tree objects only encode `path → blob`.
-///
-/// The backup file lives under `state_dir/index-snapshots/`. `Drop` removes
-/// the file best-effort once the snapshot goes out of scope.
-pub struct IndexSnapshot {
-    backup_path: PathBuf,
-}
-
-impl IndexSnapshot {
-    /// Path of the on-disk backup. Exposed for diagnostics; callers should
-    /// prefer [`restore_index`] for the actual rollback.
-    #[allow(dead_code)]
-    pub fn backup_path(&self) -> &Path {
-        &self.backup_path
-    }
-}
-
-impl Drop for IndexSnapshot {
-    fn drop(&mut self) {
-        let _ = fs::remove_file(&self.backup_path);
-    }
-}
-
-pub fn snapshot_index(ctx: &AppContext) -> Result<IndexSnapshot> {
-    snapshot_index_with_env(ctx, &[])
-}
-
-fn snapshot_index_with_env(ctx: &AppContext, envs: &[(&str, &str)]) -> Result<IndexSnapshot> {
-    let index_path = resolve_git_index_path(ctx, envs)?;
-    if !index_path.exists() {
-        return Err(anyhow!(
-            "git index missing at {}; cannot snapshot",
-            index_path.display()
-        ));
-    }
-    let snapshot_dir = ctx.state_dir.join("index-snapshots");
-    fs::create_dir_all(&snapshot_dir)
-        .with_context(|| format!("create {}", snapshot_dir.display()))?;
-    let backup_path = snapshot_dir.join(format!("snapshot-{}", uuid::Uuid::new_v4()));
-    fs::copy(&index_path, &backup_path).with_context(|| {
-        format!(
-            "back up git index from {} to {}",
-            index_path.display(),
-            backup_path.display()
-        )
-    })?;
-    Ok(IndexSnapshot { backup_path })
-}
-
-pub fn restore_index(ctx: &AppContext, snapshot: &IndexSnapshot) -> Result<()> {
-    restore_index_with_env(ctx, snapshot, &[])
-}
-
-fn restore_index_with_env(
-    ctx: &AppContext,
-    snapshot: &IndexSnapshot,
-    envs: &[(&str, &str)],
-) -> Result<()> {
-    let index_path = resolve_git_index_path(ctx, envs)?;
-    let parent = index_path
-        .parent()
-        .ok_or_else(|| anyhow!("git index path has no parent: {}", index_path.display()))?;
-    // Stage the restore next to the active index so the final rename is
-    // intra-directory (atomic on POSIX).
-    let staging = parent.join(format!(".loom-index-restore-{}", uuid::Uuid::new_v4()));
-    fs::copy(&snapshot.backup_path, &staging).with_context(|| {
-        format!(
-            "stage index restore at {} from {}",
-            staging.display(),
-            snapshot.backup_path.display()
-        )
-    })?;
-    crate::fs_util::rename_atomic(&staging, &index_path)
-        .with_context(|| format!("install restored index at {}", index_path.display()))?;
-    Ok(())
-}
-
-fn resolve_git_index_path(ctx: &AppContext, envs: &[(&str, &str)]) -> Result<PathBuf> {
-    let raw = run_git_in_with_env(&ctx.root, envs, &["rev-parse", "--git-path", "index"])?;
-    let candidate = PathBuf::from(&raw);
-    if candidate.is_absolute() {
-        Ok(candidate)
-    } else {
-        Ok(ctx.root.join(&candidate))
-    }
-}
-
-pub fn stage_path(ctx: &AppContext, path: &Path) -> Result<()> {
-    let path_str = path.to_string_lossy();
-    run_git(ctx, &["add", "--", &path_str])?;
-    Ok(())
-}
-
-pub fn commit(ctx: &AppContext, message: &str) -> Result<String> {
-    run_git(ctx, &["commit", "-m", message])?;
-    head(ctx)
-}
-
-pub fn commit_paths_if_changed(
-    ctx: &AppContext,
-    paths: &[&str],
-    message: &str,
-) -> Result<Option<String>> {
-    let paths = paths
-        .iter()
-        .filter_map(|path| match path_exists_or_is_tracked(ctx, path) {
-            Ok(true) => Some(Ok((*path).to_string())),
-            Ok(false) => None,
-            Err(err) => Some(Err(err)),
-        })
-        .collect::<Result<Vec<_>>>()?;
-
-    if paths.is_empty() {
-        return Ok(None);
-    }
-
-    for path in &paths {
-        run_git(ctx, &["add", "-A", "--", path])?;
-    }
-
-    let mut diff_args = vec!["diff", "--cached", "--quiet", "--"];
-    diff_args.extend(paths.iter().map(String::as_str));
-    let diff = run_git_allow_failure(ctx, &diff_args)?;
-    if diff.status.success() {
-        return Ok(None);
-    }
-
-    let mut commit_args = vec!["commit", "-m", message, "--"];
-    commit_args.extend(paths.iter().map(String::as_str));
-    run_git(ctx, &commit_args)?;
-    head(ctx).map(Some)
-}
-
-fn path_exists_or_is_tracked(ctx: &AppContext, path: &str) -> Result<bool> {
-    if ctx.root.join(path).exists() {
-        return Ok(true);
-    }
-
-    let output = run_git_allow_failure(ctx, &["ls-files", "--error-unmatch", "--", path])?;
-    Ok(output.status.success())
-}
-
-pub fn head(ctx: &AppContext) -> Result<String> {
-    run_git(ctx, &["rev-parse", "HEAD"])
-}
-
-pub fn short_head(ctx: &AppContext) -> Result<String> {
-    run_git(ctx, &["rev-parse", "--short", "HEAD"])
-}
-
-pub fn create_annotated_tag(ctx: &AppContext, tag: &str, message: &str) -> Result<()> {
-    run_git(ctx, &["tag", "-a", tag, "-m", message])?;
-    Ok(())
-}
-
-pub fn checkout_path_from_ref(ctx: &AppContext, reference: &str, path: &Path) -> Result<()> {
-    let path_str = path.to_string_lossy();
-    run_git(ctx, &["checkout", reference, "--", &path_str])?;
-    Ok(())
-}
-
-pub fn resolve_ref(ctx: &AppContext, reference: &str) -> Result<String> {
-    run_git(ctx, &["rev-parse", reference])
-}
-
-pub fn set_remote_origin(ctx: &AppContext, url: &str) -> Result<()> {
-    validate_git_url(url)?;
-    let has_origin = run_git_allow_failure(ctx, &["remote", "get-url", "origin"])?;
-    if has_origin.status.success() {
-        run_git(ctx, &["remote", "set-url", "origin", url])?;
-    } else {
-        run_git(ctx, &["remote", "add", "origin", url])?;
-    }
-    Ok(())
-}
-
-pub fn remote_exists(ctx: &AppContext) -> bool {
-    match run_git_allow_failure(ctx, &["remote", "get-url", "origin"]) {
-        Ok(out) => out.status.success(),
-        Err(_) => false,
-    }
-}
-
-pub fn remote_url(ctx: &AppContext) -> Result<Option<String>> {
-    let output = run_git_allow_failure(ctx, &["remote", "get-url", "origin"])?;
-    if !output.status.success() {
-        return Ok(None);
-    }
-    Ok(Some(
-        String::from_utf8_lossy(&output.stdout).trim().to_string(),
-    ))
-}
-
-pub fn fetch_origin_main_if_present(ctx: &AppContext) -> Result<bool> {
-    ensure_origin_remote_url_allowed(ctx)?;
-    let output = run_git_allow_failure(ctx, &["fetch", "origin", "main"])?;
-    if output.status.success() {
-        return Ok(true);
-    }
-
-    let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-    if stderr.contains("couldn't find remote ref main") {
-        return Ok(false);
-    }
-
-    Err(anyhow!("git fetch origin main failed: {}", stderr))
-}
-
-pub fn fetch_origin_history_branch_if_present(ctx: &AppContext) -> Result<bool> {
-    ensure_origin_remote_url_allowed(ctx)?;
-    let output = run_git_allow_failure(ctx, &["fetch", "origin", HISTORY_BRANCH])?;
-    if output.status.success() {
-        return Ok(true);
-    }
-
-    let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-    if stderr.contains("couldn't find remote ref") && stderr.contains(HISTORY_BRANCH) {
-        return Ok(false);
-    }
-
-    Err(anyhow!(
-        "git fetch origin {} failed: {}",
-        HISTORY_BRANCH,
-        stderr
-    ))
-}
-
-pub fn remote_tracking_main_exists(ctx: &AppContext) -> Result<bool> {
-    let output = run_git_allow_failure(
-        ctx,
-        &[
-            "show-ref",
-            "--verify",
-            "--quiet",
-            "refs/remotes/origin/main",
-        ],
-    )?;
-    Ok(output.status.success())
-}
-
-pub fn remote_tracking_history_exists(ctx: &AppContext) -> Result<bool> {
-    let output = run_git_allow_failure(
-        ctx,
-        &["show-ref", "--verify", "--quiet", ORIGIN_HISTORY_BRANCH_REF],
-    )?;
-    Ok(output.status.success())
-}
-
-pub fn history_branch_exists(ctx: &AppContext) -> Result<bool> {
-    let output = run_git_allow_failure(
-        ctx,
-        &["show-ref", "--verify", "--quiet", HISTORY_BRANCH_REF],
-    )?;
-    Ok(output.status.success())
-}
-
-pub fn ahead_behind_main(ctx: &AppContext) -> Result<(u32, u32)> {
-    ahead_behind_refs(ctx, "origin/main", "HEAD")
-}
-
-pub fn ahead_behind_refs(ctx: &AppContext, left: &str, right: &str) -> Result<(u32, u32)> {
-    let range = format!("{left}...{right}");
-    let output = run_git(ctx, &["rev-list", "--left-right", "--count", &range])?;
-    let mut parts = output.split_whitespace();
-    let left_only = parts
-        .next()
-        .ok_or_else(|| anyhow!("unexpected rev-list output"))?
-        .parse::<u32>()
-        .context("failed to parse left-only count")?;
-    let right_only = parts
-        .next()
-        .ok_or_else(|| anyhow!("unexpected rev-list output"))?
-        .parse::<u32>()
-        .context("failed to parse right-only count")?;
-    Ok((right_only, left_only))
-}
-
-pub fn push_main_with_tags(ctx: &AppContext) -> Result<()> {
-    ensure_origin_remote_url_allowed(ctx)?;
-    let mut args = vec!["push", "--atomic", "origin", "HEAD:main"];
-    if history_branch_exists(ctx)? {
-        args.push("loom-history:loom-history");
-    }
-    args.push("--tags");
-    run_git(ctx, &args)?;
-    Ok(())
-}
-
-pub fn pull_rebase_main(ctx: &AppContext) -> Result<()> {
-    ensure_origin_remote_url_allowed(ctx)?;
-    let output = run_git_allow_failure(ctx, &["pull", "--rebase", "origin", "main"])?;
-    if output.status.success() {
-        return Ok(());
-    }
-
-    let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-    let _ = run_git_allow_failure(ctx, &["rebase", "--abort"]);
-
-    Err(anyhow!("git pull --rebase origin main failed: {}", stderr))
-}
-
-fn ensure_origin_remote_url_allowed(ctx: &AppContext) -> Result<()> {
-    if let Some(url) = remote_url(ctx)? {
-        validate_git_url(&url)?;
-    }
-    Ok(())
-}
-
-pub fn validate_git_url(raw: &str) -> Result<()> {
-    let url = raw.trim();
-    if url.is_empty() {
-        return Err(anyhow!("git url must not be empty"));
-    }
-    if url != raw {
-        return Err(anyhow!(
-            "git url must not include leading or trailing whitespace"
-        ));
-    }
-    if url.starts_with('-') {
-        return Err(anyhow!("git url must not start with '-'"));
-    }
-    if url
-        .chars()
-        .any(|ch| ch.is_ascii_control() || ch.is_whitespace())
-    {
-        return Err(anyhow!(
-            "git url must not contain whitespace or control characters"
-        ));
-    }
-    if let Some((scheme, _rest)) = url.split_once("://") {
-        return match scheme {
-            "https" | "ssh" => Ok(()),
-            _ => Err(anyhow!(
-                "unsupported git url scheme '{}'; use https:// or ssh://",
-                scheme
-            )),
-        };
-    }
-    let path = Path::new(url);
-    if path.is_absolute() {
-        return validate_local_git_remote_path(path);
-    }
-    validate_scp_like_git_url(url)
-}
-
-fn validate_local_git_remote_path(path: &Path) -> Result<()> {
-    if !path.exists() {
-        if path.extension().is_some_and(|extension| extension == "git") {
-            return Ok(());
-        }
-        return Err(anyhow!(
-            "local git remote path does not exist: {}",
-            path.display()
-        ));
-    }
-    if path.join(".git").is_dir() {
-        return Ok(());
-    }
-    if path.join("HEAD").is_file() && path.join("objects").is_dir() {
-        return Ok(());
-    }
-    Err(anyhow!(
-        "local git remote path must point to a git repository: {}",
-        path.display()
-    ))
-}
-
-fn validate_scp_like_git_url(url: &str) -> Result<()> {
-    let Some((user_host, path)) = url.split_once(':') else {
-        return Err(anyhow!(
-            "git url must use https://, ssh://, or git@host:owner/repo.git"
-        ));
-    };
-    if user_host.is_empty() || path.is_empty() || path.starts_with(':') {
-        return Err(anyhow!(
-            "git url must use https://, ssh://, or git@host:owner/repo.git"
-        ));
-    }
-    if user_host.contains('/') || user_host == "ext" {
-        return Err(anyhow!("unsupported git url"));
-    }
-    let host = user_host
-        .rsplit_once('@')
-        .map_or(user_host, |(_, host)| host);
-    if host.is_empty() || host == "ext" {
-        return Err(anyhow!("scp-like git url must include a hostname"));
-    }
-    if host.starts_with('-') || host.starts_with('.') {
-        return Err(anyhow!("scp-like git url contains an invalid hostname"));
-    }
-    if !host
-        .chars()
-        .all(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '.' | '-' | '_'))
-    {
-        return Err(anyhow!("scp-like git url contains an invalid hostname"));
-    }
-    Ok(())
-}
-
-pub fn diff_path(ctx: &AppContext, from: &str, to: &str, path: &Path) -> Result<String> {
-    let path_str = path.to_string_lossy();
-    run_git(ctx, &["diff", from, to, "--", &path_str])
-}
-
-pub fn fsck(ctx: &AppContext) -> Result<String> {
-    run_git(ctx, &["fsck", "--no-progress"])
-}
-
-fn hash_object_file(ctx: &AppContext, path: &Path) -> Result<String> {
-    let path_str = path.to_string_lossy();
-    run_git(ctx, &["hash-object", "-w", &path_str])
-}
-
-fn hash_object_bytes(ctx: &AppContext, bytes: &[u8]) -> Result<String> {
-    run_git_in_with_input(&ctx.root, &["hash-object", "-w", "--stdin"], bytes)
-}
-
-fn read_blob(ctx: &AppContext, blob: &str) -> Result<String> {
-    run_git(ctx, &["cat-file", "-p", blob])
-}
-
-fn ensure_local_identity(ctx: &AppContext) -> Result<()> {
-    ensure_local_identity_in(&ctx.root)
-}
-
-fn ensure_local_identity_in(repo_dir: &Path) -> Result<()> {
-    if !has_local_config_in(repo_dir, "user.name")? {
-        run_git_in(repo_dir, &["config", "--local", "user.name", "loom"])?;
-    }
-    if !has_local_config_in(repo_dir, "user.email")? {
-        run_git_in(repo_dir, &["config", "--local", "user.email", "loom@local"])?;
-    }
-    Ok(())
-}
-
-fn has_local_config_in(repo_dir: &Path, key: &str) -> Result<bool> {
-    let output = run_git_allow_failure_in(repo_dir, &["config", "--local", "--get", key])?;
-    Ok(output.status.success())
-}
-
-struct TempFile {
-    path: PathBuf,
-}
-
-impl TempFile {
-    fn new(prefix: &str) -> Result<Self> {
-        let path = std::env::temp_dir().join(format!("{}-{}", prefix, uuid::Uuid::new_v4()));
-        OpenOptions::new()
-            .create_new(true)
-            .write(true)
-            .open(&path)
-            .with_context(|| format!("failed to create {}", path.display()))?;
-        Ok(Self { path })
-    }
-}
-
-impl Drop for TempFile {
-    fn drop(&mut self) {
-        let _ = fs::remove_file(&self.path);
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::state::AppContext;
+    use std::fs;
+    use std::path::Path;
     use std::process::Command;
     use uuid::Uuid;
 
@@ -775,7 +179,7 @@ mod tests {
             "expected IT-A marker in alternate index pre-snapshot, got:\n{before}"
         );
 
-        let snapshot = snapshot_index_with_env(&ctx, &envs).expect("snapshot alternate index");
+        let snapshot = index::snapshot_index_with_env(&ctx, &envs).expect("snapshot alternate index");
 
         git_ok_with_env(&dir, &["update-index", "--force-remove", "alt.txt"], &envs);
         let cleared = git_ok_with_env(&dir, &["status", "--porcelain"], &envs);
@@ -784,7 +188,7 @@ mod tests {
             "force-remove must clear alternate index entry, got:\n{cleared}"
         );
 
-        restore_index_with_env(&ctx, &snapshot, &envs).expect("restore alternate index");
+        index::restore_index_with_env(&ctx, &snapshot, &envs).expect("restore alternate index");
 
         let restored = git_ok_with_env(&dir, &["status", "--porcelain"], &envs);
         assert!(
